extract MenuIDs to separate module
