+++ app/src/main-process/menu/index.ts
@@ -1,6 +1,5 @@
 export * from './build-default-menu'
 export * from './ensure-item-ids'
 export * from './menu-event'
-export * from './menu-ids'
 export * from './crash-menu'
 export * from './get-all-menu-items'
